// src/app/modules/api-clients/src/lib/models/cmsConfig/cmsConfig-models.ts
export class LoyaltyProgram {
    tierCode?: string | undefined;
    tierName?: string | undefined;
    mainColor?: string | undefined;
    darkerColor?: string | undefined;

    init(_data?: any) {
        if (_data) {
            this.tierCode = _data["tierCode"];
            this.tierName = _data["tierName"];
            this.mainColor = _data["mainColor"];
            this.darkerColor = _data["darkerColor"];
        }
    }

    static fromJS(data: any): LoyaltyProgram {
        data = typeof data === 'object' ? data : {};
        let result = new LoyaltyProgram();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["tierCode"] = this.tierCode;
        data["tierName"] = this.tierName;
        data["mainColor"] = this.mainColor;
        data["darkerColor"] = this.darkerColor;
        return data;
    }
}

export class LoyaltyPrograms {
    items?: LoyaltyProgram[] | undefined;

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["items"])) {
                this.items = [] as any;
                for (let item of _data["items"])
                    this.items!.push(LoyaltyProgram.fromJS(item));
            }
        }
    }

    static fromJS(data: any): LoyaltyPrograms {
        data = typeof data === 'object' ? data : {};
        let result = new LoyaltyPrograms();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.items)) {
            data["items"] = [];
            for (let item of this.items)
                data["items"].push(item ? item.toJSON() : undefined as any);
        }
        return data;
    }
}